refactor: rename global.lua -> debug.lua
That was the original name of the file before I expanded it with stuff inside the loop() function. Now that the function is empty (other than calling hooks), we’re back to having only debug stuff there.
